refactor(py): async user facing api refactor
