+++ py/docs/cloud-run.md
@@ -36,7 +36,7 @@ app = Flask(__name__)
 @genkit_flask_handler(ai)
 @ai.flow()
 async def joke(name: str, ctx):
-    return await ai.agenerate(
+    return await ai.generate(
         on_chunk=ctx.send_chunk,
         prompt=f'tell a medium sized joke about {name}',
     )
