+++ py/plugins/dev-local-vector-store/src/genkit/plugins/dev_local_vector_store/local_vector_store_api.py
@@ -22,7 +22,7 @@ from abc import ABC
 from functools import cached_property
 from typing import Any
 
-from genkit.ai.veneer import Genkit
+from genkit.ai import Genkit
 from genkit.plugins.dev_local_vector_store.constant import DbValue, Params
