+++ py/samples/google-genai-imagen/src/google_genai_imagen.py
@@ -36,7 +36,7 @@ async def draw_image_with_imagen() -> str:
         The image.
     """
     config = {'number_of_images': 1, 'language': 'en'}
-    return await ai.agenerate(
+    return await ai.generate(
         prompt='Draw a cat in a hat',
         model=vertexai_name('imagegeneration@006'),
         # optional config; check README for available fields
