+++ py/samples/hello-google-genai/src/hello_google_genai.py
@@ -99,7 +99,7 @@ async def simple_generate_with_tools_flow(value: int) -> str:
     Returns:
         The generated response with a function.
     """
-    response = await ai.agenerate(
+    response = await ai.generate(
         model=googleai_name(gemini.GoogleAiVersion.GEMINI_2_0_FLASH),
         messages=[
             Message(
@@ -128,7 +128,7 @@ def gablorken_tool2(input_: GablorkenInput, ctx: ToolRunContext):
 
 @ai.flow()
 async def simple_generate_with_interrupts(value: int) -> str:
-    response1 = await ai.agenerate(
+    response1 = await ai.generate(
         model=googleai_name(gemini.GoogleAiVersion.GEMINI_2_0_FLASH),
         messages=[
             Message(
@@ -143,7 +143,7 @@ async def simple_generate_with_interrupts(value: int) -> str:
         return response1.text
 
     tr = tool_response(response1.interrupts[0], 178)
-    response = await ai.agenerate(
+    response = await ai.generate(
         model=googleai_name(gemini.GoogleAiVersion.GEMINI_2_0_FLASH),
         messages=response1.messages,
         tool_responses=[tr],
@@ -162,7 +162,7 @@ async def say_hi(name: str):
     Returns:
         The generated response with a function.
     """
-    resp = await ai.agenerate(
+    resp = await ai.generate(
         prompt=f'hi {name}',
     )
     return resp.text
@@ -179,7 +179,7 @@ async def embed_docs(docs: list[str]):
         The generated embedding.
     """
     options = {'task_type': EmbeddingTaskType.CLUSTERING}
-    return await ai.aembed(
+    return await ai.embed(
         model=googleai_name(GeminiEmbeddingModels.TEXT_EMBEDDING_004),
         documents=[Document.from_text(doc) for doc in docs],
         options=options,
@@ -196,7 +196,7 @@ async def say_hi_with_configured_temperature(data: str):
     Returns:
         The generated response with a function.
     """
-    return await ai.agenerate(
+    return await ai.generate(
         messages=[Message(role=Role.USER, content=[TextPart(text=f'hi {data}')])],
         config=GenerationCommonConfig(temperature=0.1),
     )
@@ -259,7 +259,7 @@ async def generate_character(name: str, ctx):
 
         return (await result).output
     else:
-        result = await ai.agenerate(
+        result = await ai.generate(
             prompt=f'generate an RPG character named {name}',
             output_schema=RpgCharacter,
         )
@@ -277,7 +277,7 @@ async def generate_character_unconstrained(name: str, ctx):
     Returns:
         The generated RPG character.
     """
-    result = await ai.agenerate(
+    result = await ai.generate(
         prompt=f'generate an RPG character named {name}',
         output_schema=RpgCharacter,
         output_constrained=False,
@@ -297,7 +297,7 @@ async def generate_images(name: str, ctx):
     Returns:
         The generated response with a function.
     """
-    result = await ai.agenerate(
+    result = await ai.generate(
         prompt='tell me a about the Eifel Tower with photos',
         config=GeminiConfigSchema(response_modalities=['text', 'image']),
     )
