+++ py/samples/ollama-simple-embed/src/pokemon_glossary.py
@@ -106,7 +106,7 @@ pokemon_list = [
 async def embed_pokemons() -> None:
     """Embed the Pokemons."""
     for pokemon in pokemon_list:
-        embedding_response = await ai.aembed(
+        embedding_response = await ai.embed(
             model=ollama_name(EMBEDDER_MODEL),
             documents=[Document.from_text(pokemon.description)],
         )
@@ -172,14 +172,14 @@ async def generate_response(question: str) -> GenerateResponse:
     Returns:
         A GenerateResponse object with the answer.
     """
-    input_embedding = await ai.aembed(
+    input_embedding = await ai.embed(
         model=ollama_name(EMBEDDER_MODEL),
         documents=[Document.from_text(text=question)],
     )
     nearest_pokemon = find_nearest_pokemons(input_embedding.embeddings[0].embedding)
     pokemons_context = '\n'.join(f'{pokemon.name}: {pokemon.description}' for pokemon in nearest_pokemon)
 
-    return await ai.agenerate(
+    return await ai.generate(
         model=ollama_name(GENERATE_MODEL),
         prompt=f'Given the following context on Pokemon:\n${pokemons_context}\n\nQuestion: ${question}\n\nAnswer:',
     )
