+++ py/samples/tool-interrupts/src/tool_interrupts.py
@@ -44,7 +44,7 @@ def present_questions(questions: TriviaQuestions, ctx: ToolRunContext):
 
 
 async def main() -> None:
-    response = await ai.agenerate(
+    response = await ai.generate(
         prompt='You a trivia game host. Cheerfully greet the user when they '
         + 'first join and ank them to for the theme of the trivia game, suggest '
         + "a few theme options, they don't have to use your suggestion, feel free "
@@ -57,7 +57,7 @@ async def main() -> None:
     print(response.text)
     messages = response.messages
     while True:
-        response = await ai.agenerate(
+        response = await ai.generate(
             messages=messages,
             prompt=input('Say: '),
             tools=['present_questions'],
@@ -72,7 +72,7 @@ async def main() -> None:
                 i += 1
 
             tr = tool_response(request, input('Your answer (number): '))
-            response = await ai.agenerate(
+            response = await ai.generate(
                 messages=messages,
                 tool_responses=[tr],
                 tools=['present_questions'],
