+++ py/samples/vertex-ai-imagen/src/vertex_ai_imagen_sample.py
@@ -35,7 +35,7 @@ ai = Genkit(plugins=[VertexAI()], model=vertexai_name(ImagenVersion.IMAGEN3_FAST
 async def draw_image(prompt: str):
     # config is optional
     config = ImagenOptions(number_of_images=3)
-    return await ai.agenerate(
+    return await ai.generate(
         messages=[
             Message(
                 role=Role.USER,
